Migrate the new function into the old plugin
